requirements.txt
# Core PyTorch and Transformers
torch>=2.0.0
transformers>=4.35.0
tokenizers>=0.14.0

# LoRA and Parameter-Efficient Fine-tuning
peft>=0.6.0
bitsandbytes>=0.41.0

# Data Processing
datasets>=2.14.0
numpy>=1.24.0
pandas>=2.0.0

# Evaluation and Metrics
scikit-learn>=1.3.0
rouge-score>=0.1.2
nltk>=3.8.0

# Visualization and Logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
tqdm>=4.65.0

# Configuration and Utilities
pyyaml>=6.0
argparse
json5>=0.9.0

# Optional: For advanced training features
accelerate>=0.24.0
deepspeed>=0.10.0

# Development and Testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
